requirements.txt
langchain
langchain-ollama
langchain-core
langchain-community
yfinance
gradio
pandas
numpy
requests
python-dotenv